Implemented the Confirms endpoint.
Haskell is a great language, but it definitely suffers to some extent
from some apparently historical issues. Firstly, its String classes are
a pain in the arse I have not encountered in other languages. Secondly,
its record syntax does not allow for two data types in the same module
to have fields with the same name. The latter has a fix waiting to be
merged into master, but the former seems to be a thorny problem which
the community is unwilling to resolve.