All Positions implemented, but not tested
This should have been easy, and in many ways it was, but I hadn't
sufficiently scrutinised the content of the api companion for markets
and thus hadn't noticed that the updateTimeUTC stamp was just the time
and not also the year. Which is quite irritating because it forces the
client to use things like 1970-01-01 which is wrong, and also because
you therefore cannot tell when the market was actually last updated.